tests/wpt/css-tests/css-flexbox-1_dev/xhtml1/flexbox_computedstyle_align-items-flex-end.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head><title>flexbox | computed style | align-items: flex-end</title>
<link href="http://opera.com" rel="author" title="Opera Software" />
<link href="http://www.w3.org/TR/css-flexbox-1/#propdef-align-items" rel="help" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta content="dom" name="flags" />
<style>
body {
	color: red;
	align-items: flex-end;
}
</style>
</head><body><div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("align-items"),
	          "flex-end");
});
</script>
</body></html>
